feat: remove demo daemon set
The example has been removed as we do not plan to use this with K8s.
Ensuring that the demo daemon set is working now and in the future could cause uncessary effort.
If you are interested, feel free to request that we add it again. But without third party users
the maintenance is not worth it, when we don't need it.